Fix Denial of Service
hasProof is iterating using a "var". Solidity will infer that a "unit8" is sufficient for this "var". "unit8" variables can hold values in the range 0 - 255.

If proofs grows to 256 elements, the hasProof loop will never terminate, resulting in a DoS. The notarize function is (implicitly) public, so anyone can cause this condition.

Two changes:
1. make "uint8" explicit
2. throw if someone attempts to add a 256th element to the proofs array

Not sure if you care to maintain this code, but might be nice to address in case others use your code to learn. Thanks!